+++ dbmanager.h
@@ -25,7 +25,7 @@ public:
 
     //----METHODEN FUER GUI-ANZEIGE----
     //englisch oder deutsch? qulle & zahlart können aber müssen nicht vorhanden sein, evtl. auffuellen mit null
-    bool addTransaction(QString &beschr, int betr, QString &date, QString &quelle, int kID, int bID, int zID);
+    bool addTransaction(QString &beschr, qint64 betr, QString &date, QString &quelle, int kID, int bID, int zID);
     bool showAccount();
     bool addPayment(QString &pay, int uID);
     bool delPayment(QString &pay, int uID);
